Added function to reset auth details
After Trakt's recent outage, they reset all the API keys, so the script
needs re-authenticating, but it won't do it on its own, because it wants
to use the details stored in the PropertyStore. So a function has been
added to delete the old details and force a new authentication.