Added support for backing up full liked lists
This script now backs up the full contents of liked lists, rather than
just a list of them.
Doing so involved some finagling, as liked/public lists don't quite
behave the same way as personal lists. This means that in some unknown
set of cases, despite claiming that the list is public, and has
comments and/or items, they cannot be retrieved.